修改:修改取样量设置可以输入小数位，但真实取样量为四舍五入（崔）

+++ app/src/main/java/com/wl/turbidimetric/model/GlobalConfig.kt
@@ -23,7 +23,7 @@ data class GlobalConfig(
     /**
      * 取样量
      */
-    val SamplingVolume: Int = LocalDataGlobal.Default.SamplingVolume,
+    val SamplingVolume: Double = LocalDataGlobal.Default.SamplingVolume,
     /**
      * 最后一次进入的版本号
      */
